# api/openapi.json
{"openapi":"3.0.0","info":{"title":"Cola z Kranu API","version":"1.0.0","description":"API dla platformy zgłaszania problemów z jakością wody w Polsce","contact":{"name":"API Support","email":"support@colazkreanu.pl"}},"servers":[{"url":"http://localhost:3001/api","description":"Development server"},{"url":"https://api.colazkreanu.pl/api","description":"Production server"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT","description":"Enter JWT token from /admin/auth/login"}},"schemas":{"Report":{"type":"object","properties":{"id":{"type":"integer"},"uuid":{"type":"string","format":"uuid"},"types":{"type":"array","items":{"type":"string","enum":["brown_water","bad_smell","sediment","pressure","no_water","other"]},"minItems":1,"maxItems":3},"description":{"type":"string","nullable":true},"latitude":{"type":"number","minimum":-90,"maximum":90},"longitude":{"type":"number","minimum":-180,"maximum":180},"address":{"type":"string","nullable":true},"city":{"type":"string","nullable":true},"voivodeship":{"type":"string","nullable":true},"reportedAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"status":{"type":"string","enum":["active","deleted","spam"]},"photos":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer"},"url":{"type":"string"},"mimeType":{"type":"string"},"size":{"type":"integer"}}}}}},"CreateReportRequest":{"type":"object","required":["types","latitude","longitude","reportedAt"],"properties":{"types":{"type":"array","items":{"type":"string","enum":["brown_water","bad_smell","sediment","pressure","no_water","other"]},"minItems":1,"maxItems":3},"description":{"type":"string","maxLength":2000},"latitude":{"type":"number","minimum":-90,"maximum":90},"longitude":{"type":"number","minimum":-180,"maximum":180},"address":{"type":"string"},"city":{"type":"string"},"voivodeship":{"type":"string"},"reportedAt":{"type":"string","format":"date-time"},"contactEmail":{"type":"string","format":"email"},"photos":{"type":"array","items":{"type":"string"},"maxItems":5}}},"CreateReportResponse":{"type":"object","properties":{"uuid":{"type":"string","format":"uuid"},"deleteToken":{"type":"string","format":"uuid"},"message":{"type":"string"}}},"ReportsListResponse":{"type":"object","properties":{"reports":{"type":"array","items":{"$ref":"#/components/schemas/Report"}},"total":{"type":"integer"}}},"Stats":{"type":"object","properties":{"period":{"type":"string","enum":["week","month","year","all"]},"total":{"type":"integer"},"recentCount":{"type":"integer"},"byType":{"type":"object","additionalProperties":{"type":"integer"}},"byCity":{"type":"array","items":{"type":"object","properties":{"city":{"type":"string"},"count":{"type":"integer"}}}},"byVoivodeship":{"type":"array","items":{"type":"object","properties":{"voivodeship":{"type":"string"},"count":{"type":"integer"}}}}}},"GeocodeResult":{"type":"object","properties":{"lat":{"type":"number"},"lon":{"type":"number"},"display_name":{"type":"string"}}},"GeocodeResponse":{"type":"object","properties":{"results":{"type":"array","items":{"$ref":"#/components/schemas/GeocodeResult"}}}},"HealthResponse":{"type":"object","properties":{"status":{"type":"string","enum":["healthy","unhealthy"]},"timestamp":{"type":"string","format":"date-time"},"database":{"type":"string","enum":["connected","disconnected"]}}},"Error":{"type":"object","properties":{"error":{"type":"string"},"message":{"type":"string"},"details":{"type":"array","items":{"type":"object"}}}}}},"paths":{"/admin/auth/login":{"post":{"tags":["Admin Auth"],"summary":"Admin login","description":"Authenticate admin user and receive JWT token","requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email","example":"admin@example.com"},"password":{"type":"string","format":"password","example":"securePassword123"}}}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"type":"object","properties":{"token":{"type":"string","description":"JWT token"},"admin":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"}}}}}}}},"400":{"description":"Validation error"},"401":{"description":"Invalid credentials or inactive account"}}}},"/admin/auth/me":{"get":{"tags":["Admin Auth"],"summary":"Get current admin user","description":"Get information about the currently authenticated admin user","security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Admin user information","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"integer"},"email":{"type":"string"},"name":{"type":"string"},"role":{"type":"string"},"isActive":{"type":"boolean"},"lastLoginAt":{"type":"string","format":"date-time"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}}}}},"401":{"description":"Not authenticated"}}}},"/admin/auth/change-password":{"post":{"tags":["Admin Auth"],"summary":"Change admin password","description":"Change the password for the currently authenticated admin user","security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["oldPassword","newPassword"],"properties":{"oldPassword":{"type":"string","format":"password"},"newPassword":{"type":"string","format":"password","minLength":8}}}}}},"responses":{"200":{"description":"Password changed successfully"},"400":{"description":"Validation error or invalid old password"},"401":{"description":"Not authenticated"}}}},"/admin/reports":{"get":{"summary":"List all reports (admin)","description":"Get all reports including deleted/spam for admin panel","tags":["Admin - Reports"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"limit","schema":{"type":"integer","minimum":1,"maximum":1000},"description":"Number of reports to return"},{"in":"query","name":"offset","schema":{"type":"integer","minimum":0},"description":"Number of reports to skip"},{"in":"query","name":"types","schema":{"type":"string"},"description":"Comma-separated list of report types"},{"in":"query","name":"city","schema":{"type":"string"},"description":"Filter by city"},{"in":"query","name":"status","schema":{"type":"string","enum":["active","deleted","spam"]},"description":"Filter by status"}],"responses":{"200":{"description":"List of reports","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportsListResponse"}}}},"401":{"description":"Not authenticated"}}}},"/admin/reports/{uuid}":{"get":{"summary":"Get report by UUID (admin)","description":"Get report details including deleted/spam reports","tags":["Admin - Reports"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"uuid","required":true,"schema":{"type":"string","format":"uuid"},"description":"Report UUID"}],"responses":{"200":{"description":"Report details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Report"}}}},"401":{"description":"Not authenticated"},"404":{"description":"Report not found"}}},"delete":{"summary":"Permanently delete report (admin)","description":"Hard delete a report and all associated photos","tags":["Admin - Reports"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"uuid","required":true,"schema":{"type":"string","format":"uuid"},"description":"Report UUID"}],"responses":{"200":{"description":"Report permanently deleted","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"401":{"description":"Not authenticated"},"404":{"description":"Report not found"}}}},"/admin/reports/{uuid}/status":{"patch":{"summary":"Update report status","description":"Change report status (active/deleted/spam)","tags":["Admin - Reports"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"uuid","required":true,"schema":{"type":"string","format":"uuid"},"description":"Report UUID"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["status"],"properties":{"status":{"type":"string","enum":["active","deleted","spam"]}}}}}},"responses":{"200":{"description":"Status updated successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Report"}}}},"400":{"description":"Validation error"},"401":{"description":"Not authenticated"},"404":{"description":"Report not found"}}}},"/geocode":{"get":{"summary":"Search for addresses (geocoding)","tags":["Geocode"],"parameters":[{"in":"query","name":"q","required":true,"schema":{"type":"string","minLength":2},"description":"Search query (address, city, etc.)"}],"responses":{"200":{"description":"Geocoding results","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GeocodeResponse"}}}},"400":{"description":"Validation error (query too short)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/health":{"get":{"summary":"Health check endpoint","tags":["Health"],"responses":{"200":{"description":"Service is healthy","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthResponse"}}}},"503":{"description":"Service is unhealthy","content":{"application/json":{"schema":{"allOf":[{"$ref":"#/components/schemas/HealthResponse"},{"type":"object","properties":{"error":{"type":"string"}}}]}}}}}}},"/reports":{"get":{"summary":"List all reports","tags":["Reports"],"parameters":[{"in":"query","name":"limit","schema":{"type":"integer","minimum":1,"maximum":1000},"description":"Number of reports to return"},{"in":"query","name":"offset","schema":{"type":"integer","minimum":0},"description":"Number of reports to skip"},{"in":"query","name":"types","schema":{"type":"string"},"description":"Comma-separated list of report types"},{"in":"query","name":"city","schema":{"type":"string"},"description":"Filter by city"},{"in":"query","name":"north","schema":{"type":"number"},"description":"North boundary for map filtering"},{"in":"query","name":"south","schema":{"type":"number"},"description":"South boundary for map filtering"},{"in":"query","name":"east","schema":{"type":"number"},"description":"East boundary for map filtering"},{"in":"query","name":"west","schema":{"type":"number"},"description":"West boundary for map filtering"}],"responses":{"200":{"description":"List of reports","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReportsListResponse"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"post":{"summary":"Create a new report","tags":["Reports"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReportRequest"}}}},"responses":{"201":{"description":"Report created successfully","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateReportResponse"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"429":{"description":"Rate limit exceeded","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/reports/{uuid}":{"get":{"summary":"Get report by UUID","tags":["Reports"],"parameters":[{"in":"path","name":"uuid","required":true,"schema":{"type":"string","format":"uuid"},"description":"Report UUID"}],"responses":{"200":{"description":"Report details","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Report"}}}},"404":{"description":"Report not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}},"delete":{"summary":"Delete a report","tags":["Reports"],"parameters":[{"in":"path","name":"uuid","required":true,"schema":{"type":"string","format":"uuid"},"description":"Report UUID"},{"in":"query","name":"token","required":true,"schema":{"type":"string","format":"uuid"},"description":"Delete token received when creating the report"}],"responses":{"200":{"description":"Report deleted successfully","content":{"application/json":{"schema":{"type":"object","properties":{"message":{"type":"string"}}}}}},"403":{"description":"Invalid delete token","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"404":{"description":"Report not found","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}},"410":{"description":"Delete period expired (24 hours)","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/stats":{"get":{"summary":"Get statistics","tags":["Stats"],"parameters":[{"in":"query","name":"period","schema":{"type":"string","enum":["week","month","year","all"],"default":"month"},"description":"Time period for statistics"}],"responses":{"200":{"description":"Statistics data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Stats"}}}},"400":{"description":"Validation error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}}},"tags":[]}